+++ tests/unit/test_json_utils.py
@@ -35,7 +35,7 @@ from ome_zarr_converters_tools.models import (
 def sample_tiled_image() -> TiledImage:
     acq = AcquisitionDetails(
         channels=[ChannelInfo(channel_label="DAPI")],
-        pixelsize=1.0,
+        xy_pixel_size=1.0,
         z_spacing=1.0,
         t_spacing=1.0,
     )